+++ main.go
@@ -4,12 +4,13 @@ import (
 	"bytes"
 	"flag"
 	"fmt"
-	"github.com/sue445/gitpanda/gitlab"
-	"github.com/sue445/gitpanda/webhook"
 	"net/http"
 	"os"
 	"strconv"
 	"strings"
+
+	"github.com/sue445/gitpanda/webhook"
+	"github.com/sue445/gitpanda_fetcher"
 )
 
 var (
@@ -112,12 +113,12 @@ func normalHandler(w http.ResponseWriter, r *http.Request) {
 		s := webhook.NewSlackWebhook(
 			os.Getenv("SLACK_OAUTH_ACCESS_TOKEN"),
 			os.Getenv("SLACK_VERIFICATION_TOKEN"),
-			&gitlab.URLParserParams{
-				APIEndpoint:     os.Getenv("GITLAB_API_ENDPOINT"),
-				BaseURL:         os.Getenv("GITLAB_BASE_URL"),
-				PrivateToken:    os.Getenv("GITLAB_PRIVATE_TOKEN"),
-				GitPandaVersion: Version,
-				IsDebugLogging:  isDebugLogging,
+			&fetcher.ClientParams{
+				APIEndpoint:    os.Getenv("GITLAB_API_ENDPOINT"),
+				BaseURL:        os.Getenv("GITLAB_BASE_URL"),
+				PrivateToken:   os.Getenv("GITLAB_PRIVATE_TOKEN"),
+				UserAgent:      getUserAgent(),
+				IsDebugLogging: isDebugLogging,
 			},
 		)
 		response, err := s.Request(body, truncateLines)
@@ -137,4 +138,8 @@ func normalHandler(w http.ResponseWriter, r *http.Request) {
 			panic(err)
 		}
 	}
+}
+
+func getUserAgent() string {
+	return fmt.Sprintf("gitpanda/%s (+https://github.com/sue445/gitpanda)", Version)
 }