+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## Unreleased
 [full changelog](http://github.com/sue445/gitpanda/compare/v0.14.3...main)
 
+* :bomb: **[BREAKING CHANGE]** Migrate `github.com/sue445/gitpanda/gitlab/` to `github.com/sue445/gitpanda_fetcher`
+  * https://github.com/sue445/gitpanda/pull/1972
+
 ## [v0.14.3](https://github.com/sue445/gitpanda/releases/tag/v0.14.3)
 [full changelog](http://github.com/sue445/gitpanda/compare/v0.14.2...v0.14.3)